libsensors: Integrate the quaternion sensor and use for compass
The sensor hub has a built-in integrated sensor integration scheme
which emits a unified quaternion orientation.  Use this in lieu of
Google's SensorFusion software implementation.

Also: the sensor hub magnetometer emits uncalibrated data because of
requirements on other platforms, but Android demands calibrated
numbers.  Instead of doing magnetometer calibration in the HAL, use
the orientation output to derive a faked "magenetometer" which always
points to magnetic north along the ground plane.  Call this a "leveled
compass".  This behavior is apparently kosher: future (JB 4.3+) API
versions separate the existing "calibrated" compass output from a new
"raw" one which matches the sensor hub behavior.

Finally, the leveled compass implementation requires a little rework
to the SensorIIODev class to allow for reference counting of the
stream (so that it can be active when the sensor itself is disabled).

Issue: AXIA-3283
Change-Id: I16e90fa1119f908e979ca1fa010b2dea7f136deb
Depends-Change-Id: Id4feb36f32d0033f744752e238717bcd163aebb5
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
Reviewed-on: https://otc-android.intel.com/gerrit/22344
Tested-by: jenkins autobuilder
Reviewed-by: Brian E Woodruff <brian.e.woodruff@intel.com>
Works-for-me: jenkins autobuilder
Reviewed-by: Daniel Leung <daniel.leung@intel.com>
Reviewed-by: Robert Chiras <robert.chiras@intel.com>