Sensors: Setting sensor's sensitivity value to default retrieved from fw
Currently, the sensitivity value for sensors are set to DEF_HYST_VALUE
which is set to zero. This causes a continous stream of events even
when the device is still. This patch changes the behavior by initially
reading the default value of the sensor's sensitivity from firmware and
setting it back. If we are unable to read it, then we will set it to 0.

Issue: APDEV-1304
Issue: APDEV-1000

Change-Id: I1b5b9d5e9d059b65f5a91f6867ec29e24f56d4cb
Signed-off-by: Archana Patni <archana.patni@intel.com>
Reviewed-on: https://android.intel.com/160777
Reviewed-by: Sesha, Subramony <subramony.sesha@intel.com>
Tested-by: Sesha, Subramony <subramony.sesha@intel.com>
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Persson, Per <per.persson@intel.com>
Tested-by: Persson, Per <per.persson@intel.com>
Reviewed-by: Lilja, Ola <ola.lilja@intel.com>